fix: Read replica database config not detected [DHIS2-15963]
